ex02

+++ cell04/ex02/calculator.py
@@ -3,8 +3,8 @@
 num1 = input("Give me the first number: ")
 num2 = input("Give me the second number: ")
 
-num1 = float(num1)
-num2 = float(num2)
+num1 = int(num1)
+num2 = int(num2)
 
 print("Thank you!")
